+++ pkg/vm/engine/disttae/datasource.go
@@ -23,7 +23,6 @@ import (
 	"sort"
 
 	"github.com/matrixorigin/matrixone/pkg/catalog"
-	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/common/mpool"
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
 	"github.com/matrixorigin/matrixone/pkg/container/nulls"
@@ -41,202 +40,6 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/options"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 )
-
-func UnmarshalRelationData(data []byte) (engine.RelData, error) {
-	typ := engine.RelDataType(data[0])
-	switch typ {
-	case engine.RelDataBlockList:
-		relData := buildBlockListRelationData()
-		if err := relData.UnmarshalBinary(data); err != nil {
-			return nil, err
-		}
-		return relData, nil
-	default:
-		return nil, moerr.NewInternalErrorNoCtx("unsupported relation data type")
-	}
-}
-
-var _ engine.RelData = new(blockListRelData)
-
-type blockListRelData struct {
-	//blkList[0] is a empty block info
-	//blkList []*objectio.BlockInfoInProgress
-	blklist objectio.BlockInfoSliceInProgress
-
-	//tombstones
-	tombstones engine.Tombstoner
-}
-
-func buildBlockListRelationData() *blockListRelData {
-	return &blockListRelData{
-		blklist: objectio.BlockInfoSliceInProgress{},
-	}
-}
-
-func (relData *blockListRelData) String() string {
-	var w bytes.Buffer
-	w.WriteString(fmt.Sprintf("RelData[%d]<\n", relData.GetType()))
-	if relData.blklist != nil {
-		w.WriteString(fmt.Sprintf("\tBlockList: %s\n", relData.blklist.String()))
-	} else {
-		w.WriteString("\tBlockList: nil\n")
-	}
-	if relData.tombstones != nil {
-		w.WriteString(relData.tombstones.StringWithPrefix("\t"))
-	} else {
-		w.WriteString("\tTombstones: nil\n")
-	}
-	return w.String()
-}
-
-func (relData *blockListRelData) GetShardIDList() []uint64 {
-	panic("not supported")
-}
-func (relData *blockListRelData) GetShardID(i int) uint64 {
-	panic("not supported")
-}
-func (relData *blockListRelData) SetShardID(i int, id uint64) {
-	panic("not supported")
-}
-func (relData *blockListRelData) AppendShardID(id uint64) {
-	panic("not supported")
-}
-
-func (relData *blockListRelData) GetBlockInfoSlice() objectio.BlockInfoSliceInProgress {
-	return relData.blklist.GetAllBytes()
-}
-
-func (relData *blockListRelData) GetBlockInfo(i int) objectio.BlockInfoInProgress {
-	return *relData.blklist.Get(i)
-}
-
-func (relData *blockListRelData) SetBlockInfo(i int, blk objectio.BlockInfoInProgress) {
-	relData.blklist.Set(i, &blk)
-}
-
-func (relData *blockListRelData) AppendBlockInfo(blk objectio.BlockInfoInProgress) {
-	relData.blklist.AppendBlockInfo(blk)
-}
-
-func (relData *blockListRelData) UnmarshalBinary(data []byte) (err error) {
-	typ := engine.RelDataType(types.DecodeUint8(data))
-	if typ != engine.RelDataBlockList {
-		return moerr.NewInternalErrorNoCtx("UnmarshalBinary RelDataBlockList with %v", typ)
-	}
-	data = data[1:]
-
-	sizeofblks := types.DecodeUint32(data)
-	data = data[4:]
-
-	relData.blklist = data[:sizeofblks]
-	data = data[sizeofblks:]
-
-	tombstoneLen := types.DecodeUint32(data)
-	data = data[4:]
-
-	if tombstoneLen == 0 {
-		return
-	}
-
-	relData.tombstones, err = UnmarshalTombstoneData(data[:tombstoneLen])
-	return
-}
-
-func (relData *blockListRelData) MarshalBinaryWithBuffer(w *bytes.Buffer) (err error) {
-	typ := uint8(relData.GetType())
-	if _, err = w.Write(types.EncodeUint8(&typ)); err != nil {
-		return
-	}
-
-	sizeofblks := uint32(relData.blklist.Size())
-	if _, err = w.Write(types.EncodeUint32(&sizeofblks)); err != nil {
-		return
-	}
-
-	// marshal blk list
-	if _, err = w.Write(relData.blklist); err != nil {
-		return
-	}
-
-	// marshal tombstones
-	offset := w.Len()
-	tombstoneLen := uint32(0)
-	if _, err = w.Write(types.EncodeUint32(&tombstoneLen)); err != nil {
-		return
-	}
-	if relData.tombstones != nil {
-		if err = relData.tombstones.MarshalBinaryWithBuffer(w); err != nil {
-			return
-		}
-		tombstoneLen = uint32(w.Len() - offset)
-		buf := w.Bytes()
-		copy(buf[offset:], types.EncodeUint32(&tombstoneLen))
-	}
-	return
-}
-
-func (relData *blockListRelData) GetType() engine.RelDataType {
-	return engine.RelDataBlockList
-}
-
-func (relData *blockListRelData) MarshalBinary() ([]byte, error) {
-	var w bytes.Buffer
-	if err := relData.MarshalBinaryWithBuffer(&w); err != nil {
-		return nil, err
-	}
-	buf := w.Bytes()
-	return buf, nil
-}
-
-func (relData *blockListRelData) AttachTombstones(tombstones engine.Tombstoner) error {
-	relData.tombstones = tombstones
-	return nil
-}
-
-func (relData *blockListRelData) GetTombstones() engine.Tombstoner {
-	return relData.tombstones
-}
-
-func (relData *blockListRelData) DataSlice(i, j int) engine.RelData {
-	blist := objectio.BlockInfoSliceInProgress(relData.blklist.Slice(i, j))
-	return &blockListRelData{
-		blklist:    blist,
-		tombstones: relData.tombstones,
-	}
-}
-
-func (relData *blockListRelData) GroupByPartitionNum() map[int16]engine.RelData {
-	ret := make(map[int16]engine.RelData)
-
-	blks := relData.GetBlockInfoSlice()
-	blksLen := blks.Len()
-	for idx := range blksLen {
-		blkInfo := blks.Get(idx)
-		if blkInfo.IsMemBlk() {
-			return nil
-		}
-		partitionNum := blkInfo.PartitionNum
-		if _, ok := ret[partitionNum]; !ok {
-			ret[partitionNum] = &blockListRelData{
-				tombstones: relData.tombstones,
-			}
-			ret[partitionNum].AppendBlockInfo(objectio.EmptyBlockInfoInProgress)
-		}
-		ret[partitionNum].AppendBlockInfo(*blkInfo)
-	}
-
-	return ret
-}
-
-func (relData *blockListRelData) BuildEmptyRelData() engine.RelData {
-	return &blockListRelData{
-		blklist: objectio.BlockInfoSliceInProgress{},
-	}
-}
-
-func (relData *blockListRelData) DataCnt() int {
-	return relData.blklist.Len()
-}
 
 type RemoteDataSource struct {
 	ctx  context.Context
