+++ pkg/vm/engine/disttae/datasource_test.go
@@ -35,7 +35,7 @@ func TestRelationDataV1_MarshalAndUnMarshal(t *testing.T) {
 	metaLoc := objectio.ObjectLocation(delLoc)
 	cts := types.BuildTSForTest(1, 1)
 
-	relData := buildBlockListRelationData()
+	relData := NewEmptyBlockListRelationData()
 	blkNum := 10
 	for i := 0; i < blkNum; i++ {
 		blkID := types.NewBlockidWithObjectID(objID, uint16(blkNum))
